requirements.txt
fastapi
httpx
pydantic
pydantic-settings
uvicorn[standard]
kafka-python